Create Maze Dataset class
Create Maze Dataset class and corresponding example in Jupyter notebook.
Possible areas to address are:

1) Normalization of the maze data -- currently, we divide by the L1 norm
of each row of the data matrix and multiply by a large constant.
Multiplying by a large constant seems hacky, and it would be nice to
change that.

2) If the user sets the end time to be too large, or the bin size to
be too small, this will result in an OOM error. We could fix this by
catching the error and suggesting something to the user, or in the
future we'll ideally support a "stochastic" update where only part of
the data matrix needs to be loaded in memory at once.